% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anzsic_hs}
\alias{anzsic_hs}
\title{Concordance between ANZSIC and Harmonised System Product Classification}
\format{
A data frame with 1238 rows and 2 variables:
\describe{
\item{hs_product_code}
\item{anzsic_subdivision}
}
}
\source{
\url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/5489.0Main+Features12015?OpenDocument}
}
\usage{
anzsic_hs
}
\description{
A dataset containing the relevant ANZSIC subdivision for economic complexity
product codes.
}
\keyword{datasets}
