% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{create_designs}
\alias{create_designs}
\title{Designs data}
\usage{
create_designs(year, geography = "sa2")
}
\arguments{
\item{year}{numeric. The year the designs were filed.}

\item{geography}{string. The geographic region of interest. Defaults to SA2}
}
\value{
A dataframe of designs by geography for a given year.
}
\description{
Creates a dataframe with the number of designs files in a geographic are for a given year.
}
\examples{
\dontrun{
create_designs(2016, "SA2")
}
}
