% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{create_infrastructure}
\alias{create_infrastructure}
\title{Create Infrastructure Data}
\usage{
create_infrastructure(year, geography = "sa2")
}
\arguments{
\item{year}{numeric. The year of the geographic boundaries.}

\item{geography}{string. The geographic region of interest. Defaults to SA2.}
}
\value{
A data frame with the number of university and tafe locations by a geographic area.
}
\description{
\code{create_infrastructure()} creates a data frame containing the number of universities and TAFE campuses by a geographic area for
a given year.
}
\details{
Currently only point data from 2018 is available. It is assumed that physical locations are constant between 2011 and 2016 with
the only differences arising due to changes in geographic boundaries. This will likely change in future releases if additional location
data becomes available.
}
\examples{
create_infrastructure(2011)
}
