% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_intensive_business_services.R
\name{create_kibs}
\alias{create_kibs}
\title{Create Regional Knowledge Intensive Business Services Data}
\usage{
create_kibs(year, geography = "sa2", adjust = FALSE)
}
\arguments{
\item{year}{numeric. The year to create the regional knowledge intensive business services data. Must be a census year (2011, 2016).}

\item{geography}{string. The geography level to create the regional knowledge intensive business services data. Defaults to SA2.}

\item{adjust}{logical. TRUE to remove industries classified as "not further defined"}
}
\value{
A data frame of the proportion of workers in a geographic area who are employed in knowledge intensive business service industry.
}
\description{
\code{create_kibs()} creates a data frame of the proportion of workers in a geographic area who are employed in a knowledge intensive business service industry.
}
\details{
Knowledge Intensive Business Services are defined based on the list provided in the New Zealand Government
\href{https://www.mbie.govt.nz/assets/bd287dd4f2/Knowledge-intensive-services-report.pdf}{Knowledge Intensive Services Report}
}
\examples{
create_kibs(2016)
}
