% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner_managers.R
\name{create_owner_managers}
\alias{create_owner_managers}
\title{Create Regional Owner Manager Data}
\usage{
create_owner_managers(year = 2016, geography = "sa2")
}
\arguments{
\item{year}{numeric. The year to create the regional owner manager data. Must be a census year (2011, 2016).}

\item{geography}{string. The geography level to create the regional owner manager data. Defaults to SA2.}
}
\value{
A data frame of the proportion of workers in a geographic area who are owner managers of incorporated and
unincorporated enterprises.
}
\description{
\code{create_owner_managers()} creates a data frame with the proportion of workers in a geographic area who are
owner managers of incorporated and unincorporated enterprises.
}
\details{
Data is sourced from the ABS Census of Population and Housing - SIEMP and the ABS Census of Population and Housing - EMTP

Status in employment defines an employed persons status in employment for the main job held in the week prior to Census night.
The 2016 Census replaced the employment type (EMTP) measure with the status in employment measure (SIEMP). For consistency, the
2016 SIEMP data is aggregated to match the employment types present in the 2011 Census which were:
\itemize{
\item{Employee not owning business}
\item{Owner managers of incorporated enterprises}
\item{Owner managers of unincorporated enterprises}
\item{Contributing family workers}
}
}
\examples{
create_owner_managers(2016)
}
