% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents.R
\name{create_patents}
\alias{create_patents}
\title{Patents data}
\usage{
create_patents(year, geography = "sa2")
}
\arguments{
\item{year}{numeric. The year the patents were filed.}

\item{geography}{string. The geographic region of interest. Defaults to SA2.}
}
\value{
a dataframe of patents by geography for a given year.
}
\description{
Creates a dataframe with the number of patents filed in a geographic area for a given year, and
the average number of backwards citations that a patent application included.
}
\examples{
\dontrun{
create_patents(2016, "sa2")
}
}
