% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovation.R
\name{create_regional_innovation}
\alias{create_regional_innovation}
\title{Create Regional Innovation}
\usage{
create_regional_innovation(
  year,
  geography = "sa2",
  adjust = FALSE,
  model = NULL
)
}
\arguments{
\item{year}{numeric. The year to calculate the regional innovation level. Must be a Census year (2011, 2016).}

\item{geography}{string. The geographic area to calculate regional innovation. Defaults to SA2}

\item{adjust}{logical. TRUE to remove industries and occupations classified as not further defined.}

\item{model}{string. Specify a factor analysis model.}
}
\value{
A data frame of regional innovation data with an indicator for innovation.
}
\description{
\code{create_regional_innovation()} fits a confirmatory factor analysis model to derive an indicator for innovation for a given year and
geographic area.
}
\details{
A two factor model, with a secondary factor, is confirmed.

The first factor measures the skill level of the workers of a geographic region. It is measured by
the average skill level of the region, the average level of education of the region, and the proportion of employees working
in knowledge intensive business services.

The second factor measures the patent activity happening in a region. It is measured by the number of patents, and
backward patent citations per 1000 employees in a region.

Further, it is imposed that the level of education qualification in a region is correlated with the skill level of the occupations.
}
\examples{
\dontrun{
create_regional_innovation(2011)
}
}
