% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill.R
\name{create_skill}
\alias{create_skill}
\title{Create Regional Skill Data}
\usage{
create_skill(year, geography = "sa2", adjust = FALSE)
}
\arguments{
\item{year}{numeric. The year to create the regional skill data. Must be a census year (2011, 2016).}

\item{geography}{string. The geography level to create the regional skill data. Defaults to SA2.}

\item{adjust}{logical. TRUE to remove industries classified as "not further defined"}
}
\value{
A data frame of the average skill of a geographic area for a given year.
}
\description{
\code{create_skill()} creates a data frame of the average skill level of the occupations of workers in a geographic area for a given year.
}
\details{
The Australian Skills Commission provides data on the core competencies of occupations as defined by the 2006 ANZSCO at the unit level.
Each occupation is scored across 10 core competencies, receiving a score between 1 and 10 where 10 is the most skilled.

The skill level of a region is defined as the weighted average of an occupations score and the share of employment
in that occupation in a geographic area. Skill level can range from 1 to 10, where 10 is the most skilled.
}
\examples{
create_skill(2016)
}
