% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_knowledge.R
\name{create_stem}
\alias{create_stem}
\title{Create Regional Stem Employment Data.}
\usage{
create_stem(year = 2016, geography = "sa2")
}
\arguments{
\item{year}{numeric. The year to create the regional stem employment data. Must be a census year (2011, 2016).}

\item{geography}{string. The geography level to create the regional stem employment data. Defaults to SA2.}
}
\value{
A data frame containing the proportion of workers in a geographic area who have a qualification in a STEM degree.
}
\description{
\code{create_stem()} creates a data frame with the proportion of workers in a geographic area who have a qualification in a STEM degree.
}
\details{
Data is sourced from the ABS Census of Population and Housing - QALFP.

Non-school qualifications, based on field of study describes the field of study of a person's highest completed non-school qualification.
In the Census, it counts the number of persons aged 15 years and over who stated a completed qualification.

Fields of study are defined using the 2001 version of the Australian Standard Classification of Education (ASCED).

STEM qualifications refer to any non-school qualifications in any of the following fields:
\itemize{
\item{Natural and Physical Sciences}
\item{Information Technology}
\item{Engineering and Related Technologies}
\item{Agriculture, Environmental and Related Studies}
}
}
\examples{
create_stem(2016, "SA2")
}
