% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trademarks.R
\name{create_trademarks}
\alias{create_trademarks}
\title{Trademarks data}
\usage{
create_trademarks(year, geography = "sa2")
}
\arguments{
\item{year}{numeric. The year the trademarks were filed.}

\item{geography}{string. The geographic region of interest. Defaults to SA2.}
}
\value{
dataframe of trademarks by geography for a given year.
}
\description{
Creates a dataframe with the number of trademarks filed in a geographic area for a given year.
}
\examples{
\dontrun{
create_trademarks(2016, "sa2")
}
}
