% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{designs}
\alias{designs}
\title{Intellectual Property Open Government Data Designs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114315 rows and 3 columns.
}
\usage{
designs
}
\description{
A dataset containing the location by Statistical Area Level 2 of design applications by year.
}
\details{
@format A data frame with 114315 rows and 3 variables:
\descibe{
\item{application_id}
\item{sa2_name}{name, in 2016 boundaries}
\item{year}
}

@source \url{https://data.gov.au/data/dataset/intellectual-property-government-open-data-2019}
}
\keyword{datasets}
