% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igo.R
\name{igo}
\alias{igo}
\title{Industrial Growth Opportunities}
\usage{
igo(
  year,
  region = NULL,
  product = NULL,
  .export_value_limit = 0,
  .cog_limit = 0,
  .rca_limit = 1,
  .ica_limit = 1,
  ...
)
}
\arguments{
\item{year}{numeric. Which year to calculate the industrial growth opportunities.}

\item{region}{character. Show only industrial growth opportunities for a given region. The default (NULL) will return
industrial growth opportunities for all regions.}

\item{product}{character. Show only regions with an opportunitiy in a given product. The default (NULL) will return
industrial growth opportunities for all products}

\item{.export_value_limit}{(optional) numeric. The minimum state export value for a product required to be considered
a growth opportunity for a region. The default (0) excludes products which were not exported.}

\item{.cog_limit}{(optional) numeric. The minimum benefit that a product will generate for a region. The default (0) excludes all
products with negative benefits.}

\item{.rca_limit}{(optional) numeric. The maximum value for the level of revealed comparative advantage in the production of an
opportunity in a state. The default (1) excludes all products which are already produced with comaprative advantage.}

\item{.ica_limit}{(optional) numeric. The minimum value for the level of industrial comparative advantage. The default (1) excludes
all products which are made in industries in which the region does not have a higher proportion of employment than Australia.}

\item{...}{(optional). Additional arguments passed to \code{ica()}}
}
\value{
tibble
}
\description{
Industrial Growth Opportunities
}
