% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kibs_industry_class}
\alias{kibs_industry_class}
\title{Knowledge intensive business services industries by ANZSIC class.}
\format{
An object of class \code{character} of length 68.
}
\usage{
kibs_industry_class
}
\description{
A list of the 68 ANZSIC industry classes considered to be knowledge intensive businesses services (KIBS).
Knowledge Intensive Business Services are defined based on the list provided in the New Zealand Government
\href{https://www.mbie.govt.nz/assets/bd287dd4f2/Knowledge-intensive-services-report.pdf}{Knowledge Intensive Services Report}
}
\keyword{datasets}
