% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patents}
\alias{patents}
\title{Intellectual Property Open Government Data Patents}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 385646 rows and 4 columns.
}
\source{
\url{https://data.gov.au/data/dataset/intellectual-property-government-open-data-2019}
}
\usage{
patents
}
\description{
A dataset containing the location by Statistical Area Level 2 of patent applications by year.
}
\details{
@format A data frame with 385646 rows and 4 variables:
\descibe{
\item{australian_appl_no}
\item{sa2_name}{name, in 2016 boundaries}
\item{year}
\item{primary_ipc_mark_value}{patent category}
}
}
\keyword{datasets}
