% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_data}
\alias{product_data}
\title{Economic complexity product data}
\format{
A data frame with 6406 products and 5 variables:
\describe{
\item{product_id}
\item{hs_product_code}{product code, harmonised system 1992}
\item{hs_product_name_short_en}{short english name of the product}
\item{level}
\item{parent_id}
}
}
\source{
\url{https://dataverse.harvard.edu/dataverse/atlas}
}
\usage{
product_data
}
\description{
A dataset containing the name and level of economic complexity products.
}
\keyword{datasets}
