% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_2011}
\alias{sa2_2011}
\title{Simple features geometry of 2011 SA2 boundaries}
\format{
A simple feature collection with 2214 features and 14 fields:
\describe{
\item{sa2_code}
\item{sa2_5dig}
\item{sa2_name}
\item{sa3_code}
\item{sa3_name}
\item{sa4_code}
\item{sa4_name}
\item{gcc_code}
\item{gcc_name}
\item{state_code}
\item{state_name}
\item{areasqkm}{area of the sa2 region, square km}
\item{cent_long}{longitude of the centroid}
\item{cent_lat}{lattitude of the centroid}
\item{geometry}{sf}
}
}
\source{
\url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/1270.0.55.001July\%202011}
}
\usage{
sa2_2011
}
\description{
A dataset containing all standard ABS geographies for 2011.
}
\keyword{datasets}
