% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_2016_to_sa2_2011}
\alias{sa2_2016_to_sa2_2011}
\title{Concordance between 2016 and 2011 SA2 boundaries}
\format{
A data frame with 2377 rows and 3 variables:
\describe{
\item{sa2_name_2016}
\item{sa2_name_2011}
\item{ratio}{the proportion of the 2011 SA2 boundary in the 2016 boundary}
}
}
\usage{
sa2_2016_to_sa2_2011
}
\description{
A dataset containing the concordance between SA2 boundaries for 2016 and 2011.
}
\keyword{datasets}
