% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_indp2_2011}
\alias{sa2_indp2_2011}
\title{Employment by industry of employment by place of work for 2011}
\format{
A data frame with 232470 rows and 4 variables:
\describe{
\item{sa2_name}
\item{industry}{industry of employment, subdivision}
\item{employment}{number of employees, place of work}
\item{year}
}
}
\source{
ABS Census TableBuilder Pro
}
\usage{
sa2_indp2_2011
}
\description{
A dataset containing the number of employees per ANZSIC subdivision who worked in a SA2 region in 2011.
}
\keyword{datasets}
