% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_indp4_2016}
\alias{sa2_indp4_2016}
\title{Employment by industry by place of work for 2016}
\format{
A data frame with 240660 rows and 4 variables:
\describe{
\item{sa2_name}
\item{industry}{industry of employment, anzsic class}
\item{employment}{number of employees, place of work}
\item{year}
}
}
\source{
ABS Census TableBuilder Pro
}
\usage{
sa2_indp4_2016
}
\description{
A dataset containing the number of employees per ANZSIC class who worked in a SA2 region in 2011.
}
\keyword{datasets}
