% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_occp4_2011}
\alias{sa2_occp4_2011}
\title{Employment by occupation by place of work for 2011}
\format{
A data frame with 1040904 rows and 4 variables:
\describe{
\item{sa2_name}
\item{anzsco_name}{occupation of employment, anzsco unit}
\item{employment}{number of employees, place of work}
\item{year}
}
}
\source{
ABS Census TableBuilder Pro
}
\usage{
sa2_occp4_2011
}
\description{
A dataset containing the number of employees per ANZSCO unit who worked in a SA2 region in 2011.
}
\keyword{datasets}
