% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_occp4_2016}
\alias{sa2_occp4_2016}
\title{Employment by occupation by place of work for 2016}
\format{
A data frame with 1086407 rows and 4 variables:
\describe{
\item{sa2_name}
\item{anzsco_name}{occupation of employment, anzsco unit}
\item{employment}{number of employees, place of work}
\item{year}
}
}
\source{
ABS Census TableBuilder Pro
}
\usage{
sa2_occp4_2016
}
\description{
A dataset containing the number of employees per ANZSCO unit who worked in a SA2 region in 2016.
}
\keyword{datasets}
