% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_qallp1_2016}
\alias{sa2_qallp1_2016}
\title{Employment by education level by place of work for 2016}
\format{
A data frame with 16044 rows and 4 variables:
\describe{
\item{sa2_name}
\item{education_level}{education level of employment}
\item{employment}{number of employees, place of work}
\item{year}
}
}
\source{
ABS Census TableBuilder Pro
}
\usage{
sa2_qallp1_2016
}
\description{
A dataset containing the number of employees per highest education level who worked in a SA2 region in 2016.
}
\keyword{datasets}
