% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_economic_complexity}
\alias{state_economic_complexity}
\title{Australian States economic complexity data}
\format{
A data frame with 248496 rows and 13 variables:
\describe{
\item{location_code}{state name}
\item{hs_product_code}{product code, harmonised system 1992}
\item{export_value}{value of product export, in millions USD}
\item{year}
\item{diversity}{number of products exported with comparative advantage}
\item{ubiquity}{number of countries which export the product with comparative advantage}
\item{mcp}{indicator of comparative advantage, 1 if rca >=1}
\item{eci}{economic complexity index}
\item{pci}{product complexity index}
\item{density}{proportion of related products with comparative advantage}
\item{coi}{complexity outlook index}
\item{cog}{complexity outlook gain}
\item{rca}{revealed comparative advantage}
}
}
\source{
\url{https://atlas.cid.harvard.edu/glossary}
}
\usage{
state_economic_complexity
}
\description{
A dataset containing economic complexity indicators for Australian States between 1996 and 2020.
}
\keyword{datasets}
