% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trademarks}
\alias{trademarks}
\title{Intellectual Property Open Government Data Trademarks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1121886 rows and 3 columns.
}
\source{
\url{https://data.gov.au/data/dataset/intellectual-property-government-open-data-2019}
}
\usage{
trademarks
}
\description{
A dataset containing the location by Statistical Area Level 2 of trademark applications by year.
}
\details{
@format A data frame with 1121886 rows and 3 variables:
\descibe{
\item{tm_number}
\item{sa2_name}{name, in 2016 boundaries}
\item{year}
}
}
\keyword{datasets}
