% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{create_plants}
\alias{create_plants}
\title{Plant Breeder Rights data}
\usage{
create_plants(year, geography = NULL)
}
\arguments{
\item{year}{numeric. The year the plant breeder right application was submitted.}

\item{geography}{string. The geographic region of interest. Defaults to SA2}
}
\value{
A dataframe of plant breeder rights by geography for a given year.
}
\description{
Creates a datafrmae with the number of plant breeder rights applied for in a geographic area for a given year.
}
\examples{
\dontrun{
create_plants(2016, "SA2")
}
}
