% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{ica}
\alias{ica}
\title{Industrial Comparative Advantage}
\usage{
ica(years = "all", ...)
}
\arguments{
\item{years}{Which year to calculate the industrial comparative advantage. The default, "all", will calculate the
industrial comparative advantage for all available years. Available years are 2011 and 2016.}

\item{...}{Additional options passed to the calculation of the location quotient. See ?location_quotient for more information.}
}
\value{

}
\description{
Industrial Comparative Advantage
}
