% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igo.R
\name{igo}
\alias{igo}
\title{Industrial Growth Opportunities}
\usage{
igo(
  year,
  region = NULL,
  product = NULL,
  .export_value_limit = 0,
  .cog_limit = 0,
  .rca_limit = 1,
  .ica_limit = 1,
  ...
)
}
\arguments{
\item{year}{numeric. Which year to calculate the industrial growth opportunities.}

\item{region}{character. Show only industrial growth opportunities for a given region. The default (NULL) will return
industrial growth opportunities for all regions.}

\item{product}{character. Show only regions with an opportunity in a given product. The default (NULL) will return
industrial growth opportunities for all products}

\item{.export_value_limit}{(optional) numeric. The minimum state export value for a product required to be considered
a growth opportunity for a region. The default (0) excludes products which were not exported.}

\item{.cog_limit}{(optional) numeric. The minimum benefit that a product will generate for a region. The default (0) excludes all
products with negative benefits.}

\item{.rca_limit}{(optional) numeric. The maximum value for the level of revealed comparative advantage in the production of an
opportunity in a state. The default (1) excludes all products which are already produced with comparative advantage.}

\item{.ica_limit}{(optional) numeric. The minimum value for the level of industrial comparative advantage. The default (1) excludes
all products which are made in industries in which the region does not have a higher proportion of employment than Australia.}

\item{...}{(optional). Additional arguments passed to \code{ica()}.}
}
\value{
a data frame of industrial growth opportunities.
}
\description{
\code{igo()} creates a data frame of the industrial growth opportunities across Statistical Area Level 2 regions for 2011 and 2016.
}
\details{
The industrial growth opportunities are derived from the industrial comparative advantage data (see \code{ica()}), and a State based
model of economic complexity, developed by the Australian Industrial Transformation Institute at Flinders University,
in collaboration with the Government of South Australia.

Industrial growth opportunities capture, at the product level, what industrial development would be:
\itemize{
\item most beneficial for a region, and
\item most suitable for a region based on its industrial strengths.
}
}
\examples{
igo(2016)

# Industrial growth opportunities for a specific region
igo(2016, region = "Adelaide")

# Specifying a product may help to identify a region best suited to develop a new industry
igo(2016, product = "Artificial graphite")

# Override defaults.
# Increasing the export value limit (from 0) focuses the opportunities
# on those with a stronger existing industrial base.

# Increasing the COG limit (from 0) focuses the opportunities
# on those which would bring a higher benefit to the region.

# Decreasing the RCA limit (from 1) focuses the opportunities
# on those which are less developed in the State.

# Increasing the ICA limit (from 1) focuses the opportunities
# on industries in which a region has a significant comparative advantage.

igo(2016, .export_value_limit = 1000, .cog_limit = 0.5, .rca_limit = 0.5, .ica_limit = 0.5)


}
