% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_quotient.R
\name{location_quotient}
\alias{location_quotient}
\title{Location Quotient}
\usage{
location_quotient(
  data,
  options = list(min_value = 150, "total", "x", "y", "value")
)
}
\arguments{
\item{data}{a dataframe}

\item{options}{a list of options.
Available options include "min_value", "x", "y", and "value".
"min_value" specifies the smallest total value allowable in a geographic
region.
"x" specifies the name of the x variable in the dataframe
"y" specifies the name of the y variable in the dataframe
"value" specifies the name of the value variable in the dataframe}
}
\value{
tibble
}
\description{
Calculate location quotient from any data which contains a column
specifying a geographic variable, a second variable (???) and a value
(this is not written well).
}
\examples{
\dontrun{
location_quotient(sa2_indp2, options = list("min_value" = 100))
}
}
