% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_quotient.R
\name{location_quotient}
\alias{location_quotient}
\title{Location Quotient}
\usage{
location_quotient(
  data,
  min_value = 0,
  total_var = NULL,
  geography = NULL,
  y_var = NULL,
  value_var = NULL
)
}
\arguments{
\item{data}{a data frame or data frame like object.}

\item{min_value}{(numeric). The minimum value of \code{total_var} for inclusion in the location quotient calculation.}

\item{total_var}{(string). The name of the variable whose values must be greater than the \code{min_value}.}

\item{geography}{(string). The region to calculate the location quotient.}

\item{y_var}{(string). The name of the non geography variable in the data. Defaults to 'industry'.}

\item{value_var}{(string).  The name of the variable which contains the values for the geography and \code{y_var} pairs. Defaults to employment.}
}
\value{

}
\description{
\verb{location quotient()} calculates the location quotient for any data for geography based data, such as
revealed comparative advantage (RCA) from country-product-export data, or industrial comparative advantage (ICA) from
region-industry-employment data.
}
\details{
By default, assumes the input data contains region-industry-employment data.
}
\examples{
location_quotient(sa2_indp_2011)

}
