% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_innovation}
\alias{regional_innovation}
\title{Innovation metrics of SA2 regions for 2011 and 2016.}
\format{
A data frame with 4325 rows and 20 variables:
\describe{
\item{sa2_name}{name of the sa2 region. For 2011, uses the 2011 Australian Statistical Geography Standard.
For 2016, uses the 2016 Australian Statistical Geography Standard}
\item{sa2_code}{8-digit code assigned to the sa2 name}
\item{year}{}
\item{patents}{number of patents per 1000 workers}
\item{backwards_citations}{number of backward citations}
\item{designs}{number of designs per 1000 workers}
\item{trademarks}{number of trademarks per 1000 workers}
\item{plants}{number of plant breeder rights applications per 1000 workers}
\item{kibs}{proportion of workers in knowledge intensive business services}
\item{skill}{average skill level of region occupations}
\item{owner_managers}{proportion of owner-managers}
\item{unis}{number of universities}
\item{tafes}{number of tafes}
\item{qualification}{average education level of workers}
\item{stem}{proportion of workers with a stem qualification}
\item{employment}{total employment}
\item{infrastructure}{number of universities and tafes}
\item{human_knowledge}{factor derived from kibs, skill, and qualification}
\item{patent_output}{factor derived from patents and backwards citations}
\item{innovation}{second-order factor derived from human_knowledge and patent_output}
}
}
\usage{
regional_innovation
}
\description{
A dataset containing a measure of innovation, inputs to innovation, and other innovation
metrics for SA2 regions.
}
\keyword{datasets}
